refactor: rename events managers for progress and history views
