+++ src/historyView/HistoryViewContentProvider.ts
@@ -18,7 +18,10 @@ export class HistoryViewContentProvider extends BaseViewContentProvider {
       // History view specific modules
       domHandlersUri: this.getWebviewUri(webview, 'modules/domHandlers.js'),
       constantsUri: this.getWebviewUri(webview, 'modules/constants.js'),
-      eventsUri: this.getWebviewUri(webview, 'modules/uiManagers/Events.js'),
+      eventsUri: this.getWebviewUri(
+        webview,
+        'modules/uiManagers/HistoryEvents.js',
+      ),
       historyRendererUri: this.getWebviewUri(
         webview,
         'modules/uiManagers/HistoryRenderer.js',
