+++ src/historyView/index.html
@@ -20,7 +20,7 @@
         "imports": {
           "./modules/domHandlers.js": "${domHandlersUri}",
           "./modules/constants.js": "${constantsUri}",
-          "./modules/uiManagers/Events.js": "${eventsUri}",
+          "./modules/uiManagers/HistoryEvents.js": "${eventsUri}",
           "./modules/uiManagers/HistoryRenderer.js": "${historyRendererUri}",
           "./modules/historyViewState.js": "${historyViewStateUri}",
           "./modules/uiManagers/SearchManager.js": "${searchManagerUri}",
