+++ src/progressView/index.html
@@ -35,7 +35,7 @@
           "./modules/uiManagers/Toolbar.js": "${toolbarUri}",
           "./modules/uiManagers/Status.js": "${statusUri}",
           "./modules/uiManagers/FileList.js": "${fileListUri}",
-          "./modules/uiManagers/Events.js": "${eventsUri}",
+          "./modules/uiManagers/ProgressEvents.js": "${eventsUri}",
           "./modules/katexMacros.js": "${katexMacrosUri}",
           "@common/webviewContext.js": "${webviewContextUri}",
           "@common/webview/commands.js": "${commandsUri}",
